feat(api): enforce rate limits on LLM, PDF, and auth routes
The 385-line rate limiter defined per-endpoint limits but was never
registered as middleware, leaving LLM endpoints (real Anthropic spend)
unthrottled. The middleware is now registered, enforced only on the
static expensive routes in RATE_LIMITS (no catch-all default), with a
RATE_LIMIT_ENABLED kill switch (off in the test suite). Config and the
unconsumed quota block split into rate_limit_config.py/usage_quotas.py
per the 300-line rule. In-memory storage x2 workers means effective
limits up to 2x configured — acceptable for MVP; Redis path remains.
