+++ app/main.py
@@ -10,6 +10,7 @@ import os
 from app.api.v1.router import api_router
 from app.core.config import settings
 from app.core.database import Base, db, init_db
+from app.middleware.rate_limiter import rate_limit_middleware
 
 # Configure logging
 logging.basicConfig(level=logging.INFO)
@@ -50,6 +51,9 @@ app.add_middleware(
     allow_methods=["*"],
     allow_headers=["*"],
 )
+
+# Rate limiting on expensive routes (LLM, PDF, auth) — see RATE_LIMITS
+app.middleware("http")(rate_limit_middleware)
 
 # Include API routes
 app.include_router(api_router, prefix="/api/v1")
