+++ app/middleware/__init__.py
@@ -1,4 +1,5 @@
 """Middleware package"""
-from .rate_limiter import rate_limiter, rate_limit_middleware, get_token_limit
+from .rate_limit_config import get_token_limit
+from .rate_limiter import rate_limiter, rate_limit_middleware
 
 __all__ = ["rate_limiter", "rate_limit_middleware", "get_token_limit"]
